Load the correct fib.so file.
The file si/fib.so is not the same as tmp.si/fib.so. For one, Bigloo is
now case-sensitive, so the file in tmp.si/ defines the global ``fib'',
whereas the file in si/ defines ``FIB''. More importantly, the file in
si/fib.so is missing a bytecode, SHALLOW-ARGUMENT-REF1 (bytecode 2),
which is the sith bytecode from the start (zero-indexed) in
tmp.si/fib.so.

Attempting to load si/fib.so in tmp.si/after.scm was resulting in a
runtime execption when attempting to call CALL2-= with a boolean #f in
the *arg1* parameter. The exception was causing test.chap7g to fail.