Don't try to DEFINE toplevel procedures in macro expansion.
Gambit doesn't bind the symbols in the toplevel env, but rather in the
lexical env of the macro.

Bind the symbols to wrong initially and SET! them to the correct
definition at the end of the expansion of define-instruction-set.

Fixes test.chap7{d,e,g,h,i} and test.chap8{d,h,i} and evalf.test.chap8f.