Initialize vars with create-number, not raw scheme integers.
This allows the user to drop into (chapter4-interpreter), and
immediately run (set! foo 42) or similar without getting an error.

The previous version would generate a type error in transcode-back,
since new2-evaluate-set! returns the previous value of the variable
being set, which in this case was a regular scheme int, 0, which caused
transcode-back to choke on the value when it attempted to send it the
'type message.