Increase the benchmark prime range from 50 to 500.
Even the slowest chap5a interpreter was reporting 0 seconds of elapsed
time with a max of 50. Increasing the prime range to 500 results in
runtimes on my ~4 year old laptop for bench.chap5a that are comparable
to the times listed at the bottom of chap5-bench.scm, which are probably
from ~20 years ago.

At any rate, increasing the range to 500 is sufficient to show a large
difference between the runtime for bench.chap5a (~53 seconds) and
bench.chap6a (~12 seconds).