Remove obsolete test.chap10l target.
The test.chap10l target is meant to test chap10k by direct
interpretation (rather than generating C). The interpreter code has
bit-rotted and no longer works. The root of the problem is that the
files chap10{e,h,k} etc. mix code that's only needed for generating C
with code that's needed for generic program transformations (e.g. CPS)
and which is therefore needed by the interpreter. Since the Make target
was marked as obsolete by the original author, I'm removing it rather
than trying to fix the interpreter.