Update test.chap4 for guile 2.2.1
As of guile 2.2.1, all literal constants are read-only. Attempting to
set-cdr! on a quoted list, like (set-cdr! '(1 2 3) '()), is no longer
supported and will throw an error.

Specifically, src/chap4.tst was failing with the following error:

    ERROR: In procedure set-cdr!: Wrong type argument in position
    1 (expecting mutable pair): (#\e #\i #\o #\u)

In the case of the chap4 tests, the affected functions can all be
switched to calling `list` to allocate the lists, rather than using
quoted list literals.

See the guile 2.2.1 release notes for more info:
    https://lists.gnu.org/archive/html/guile-devel/2017-04/msg00034.html