Add test.release and test.all targets
The test.release target is a convenience for doing release testing. It
runs test.interpreters and then test.all for all valid schemes. If
test.release passes, then all tests are passing for all schemes.

The test.all target is like the grand.test target, but it exits with
failure as soon as a failing test is encountered, rather than
continuing to run the rest of the tests in the suite. The test.all
target also doesn't spew all test output on stdout/stderr, just
progress messages about what tests it's running. The plan is to
eventually expand the test.all target to include every test in the
Makefile, not just the ones in grand.test.